Refactor useful custom relaxation methods into ssc/relaxation.py
