+++ ssc/__init__.py
@@ -1,2 +1,3 @@
 from ssc.ssc import SSC         # noqa: F401
 from ssc.patch import PatchPC   # noqa: F401
+from ssc.relaxation import *    # noqa: F401
